metlife_agents_backend/pyproject.toml
[project]
name = "metlife-agents-backend"
version = "0.1.0"
description = "MetLife agents backend service"
readme = "metlife_agents_backend/README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "langgraph>=1.1.6",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.0.0",
    "psycopg[binary]>=3.2.9",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.44.0",
    "pyjwt>=2.8.0",
    "alembic>=1.14.0",
    "bcrypt>=5.0.0",
    "langchain>=1.2.15",
    "openai>=2.32.0",
    "sse-starlette>=3.3.4",
    "langchain-openai>=0.3.0",
    "langgraph-checkpoint-postgres==3.0.5",
    "langgraph-checkpoint-sqlite==3.0.3",
    "psycopg-pool==3.3.0",
]
